TST: Refactored test_split_grid.py
